Move Core\BaseRouter and Core\BaseView to namespace Core\Base
